api/requirements.txt
binance==0.3
fastapi==0.94.0
pandas==1.5.2
python-dotenv==1.0.0
python_binance==1.0.16
python_dateutil==2.8.2
SQLAlchemy==1.4.46
SQLAlchemy_Utils==0.40.0
uvicorn==0.21.0
PyMySQL==1.0.2
cryptography==39.0.0
python-dotenv==1.0.0
statsmodels==0.14.0